data_scorer/llm_as_judge/requirements.txt
openai
pydantic
pyyaml
rich
typer
aiofiles
pandas
orjson 